[refactor] Refactor the package name
パッケージ名をリファクタ
