+++ uilogicImpl/home/src/commonMain/kotlin/tech/takahana/iconwallpapaer/uilogic/home/HomeUiLogicImpl.kt
@@ -2,7 +2,7 @@ package tech.takahana.iconwallpapaer.uilogic.home
 
 import kotlinx.coroutines.flow.SharedFlow
 import tech.takahana.iconwallpaper.uilogic.home.HomeUiLogic
-import tech.takahana.iconwallpaper.usecase.onboarding.HomeUseCase
+import tech.takahana.iconwallpaper.usecase.home.HomeUseCase
 
 class HomeUiLogicImpl(
     private val useCase: HomeUseCase
